python_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-multipart==0.0.6
python-jose==3.3.0
sqlite-utils==3.35
aiosqlite==0.19.0
psycopg2-binary==2.9.9
python-escpos==3.0a8
python-escpos==3.0a8

